feat: separate CrimeMap and PredictionMap
