--- src/modules/map/components/CrimeMap/crimeMap.scss
@@ -1,13 +1,0 @@
-.crimeInfo {
-  &__title {
-    margin: 0 0 0.5rem;
-  }
-  &__description {
-    margin: 0 0 0.5rem;
-  }
-  &__item {
-    .itemPercentage {
-      margin: 0 0.25rem 0 0.5rem;
-    }
-  }
-}
